mejoras segun el sonar
