+++ src/main/java/work/appDeploys/equipmentControlSystem/EquipmentControlSystemApplication.java
@@ -1,0 +1,13 @@
+package work.appDeploys.equipmentControlSystem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EquipmentControlSystemApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EquipmentControlSystemApplication.class, args);
+    }
+
+}
