+++ src/main/java/work/appDeploys/equipmentControlSystem/constants/MessageResource.java
@@ -1,0 +1,36 @@
+package work.appDeploys.equipmentControlSystem.constants;
+
+public class MessageResource {
+
+    public static final  String SCHOOL_EXIST_NOT_SAVE= "The School already exist, cant be save";
+
+    public static final  String SCHOOL_NOT_EXIST_NOT_DELETE= "The School no exist, cant be delete";
+
+    public static final  String SCHOOL_NOT_EXIST_NOT_UPDATE= "The School no exist, cant be update";
+
+    public static final  String SCHOOL_EXIST_NAME_NOT_UPDATE= "The School name already exist, cant be update";
+
+    public static final  String USER_BAT_EMAIL = "You entered an incorrect email address, cant be save";
+
+    public static final  String USER_BAT_PASSWORD = "You entered an incorrect password" +
+            "the password must contain: " +
+            "1. start of the string" +
+            "2. a digit must occur at least once" +
+            "3. a lower case letter must occur at least once" +
+            "4. an upper case letter must occur at least once" +
+            "5. 4-8 character password, both inclusive" +
+            "6.  end of the string" +
+            ", cant be save";
+
+    public static final  String USER_BAT_ROL = "Bat rol, cant be save";
+
+    public static final  String USERS_EXIST_NOT_SAVE= "The user already exist, cant be save";
+
+    public static final  String USERS_EXIST_NOT_DELETE= "The user no exist, cant be delete";
+
+    public static final  String USERS_NOT_EXIST_NOT_UPDATE= "The user no exist, cant be update";
+
+    public static final  String USERS_EMAIL_ALREADY_EXIST_NOT_UPDATE= "The email alredy exist in the other user, cant be update";
+
+    public static final String UPDATE_FAIL = ", cant by update";
+}
