+++ src/main/java/work/appDeploys/equipmentControlSystem/exceptions/SchoolExceptionBadRequest.java
@@ -1,0 +1,11 @@
+package work.appDeploys.equipmentControlSystem.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class SchoolExceptionBadRequest extends RuntimeException {
+    public SchoolExceptionBadRequest(String message) {
+        super(message);
+    }
+}
