+++ src/main/java/work/appDeploys/equipmentControlSystem/exceptions/UsersExceptionBadRequest.java
@@ -1,0 +1,11 @@
+package work.appDeploys.equipmentControlSystem.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class UsersExceptionBadRequest extends RuntimeException {
+    public UsersExceptionBadRequest(String message) {
+        super(message);
+    }
+}
