+++ src/main/java/work/appDeploys/equipmentControlSystem/models/Users.java
@@ -1,0 +1,23 @@
+package work.appDeploys.equipmentControlSystem.models;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table
+@Data
+public class Users {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String email;
+    private String password;
+    private String name;
+    private String rol;
+
+}
