+++ src/main/java/work/appDeploys/equipmentControlSystem/repositories/SchoolRepository.java
@@ -1,0 +1,13 @@
+package work.appDeploys.equipmentControlSystem.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import work.appDeploys.equipmentControlSystem.models.School;
+
+import java.util.Optional;
+
+@Repository
+public interface SchoolRepository extends JpaRepository<School, Long> {
+    Optional<School> findByName(String name);
+
+}
