+++ src/main/java/work/appDeploys/equipmentControlSystem/repositories/UsersRepository.java
@@ -1,0 +1,13 @@
+package work.appDeploys.equipmentControlSystem.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import work.appDeploys.equipmentControlSystem.models.Users;
+
+import java.util.Optional;
+
+@Repository
+public interface UsersRepository extends JpaRepository<Users, Long> {
+    Optional<Users> findByEmail(String email);
+    Optional<Users> findByName(String name);
+}
