--- src/main/java/work/appdeploys/equipmentControlSystem/EquipmentControlSystemApplication.java
@@ -1,13 +1,0 @@
-package work.appdeploys.equipmentControlSystem;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EquipmentControlSystemApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EquipmentControlSystemApplication.class, args);
-    }
-
-}
