--- src/main/java/work/appdeploys/equipmentControlSystem/exceptions/SchoolExceptionBadRequest.java
@@ -1,11 +1,0 @@
-package work.appdeploys.equipmentControlSystem.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class SchoolExceptionBadRequest extends RuntimeException {
-    public SchoolExceptionBadRequest(String message) {
-        super(message);
-    }
-}
