--- src/main/java/work/appdeploys/equipmentControlSystem/exceptions/SchoolExceptionInternalError.java
@@ -1,11 +1,0 @@
-package work.appdeploys.equipmentControlSystem.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-public class SchoolExceptionInternalError extends RuntimeException {
-    public SchoolExceptionInternalError(String message) {
-        super(message);
-    }
-}
