--- src/main/java/work/appdeploys/equipmentControlSystem/repositories/SchoolRepository.java
@@ -1,13 +1,0 @@
-package work.appdeploys.equipmentControlSystem.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import work.appdeploys.equipmentControlSystem.models.School;
-
-import java.util.Optional;
-
-@Repository
-public interface SchoolRepository extends JpaRepository<School, Long> {
-    Optional<School> findByName(String name);
-
-}
