--- src/main/java/work/appdeploys/equipmentControlSystem/repositories/UsersRepository.java
@@ -1,13 +1,0 @@
-package work.appdeploys.equipmentControlSystem.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import work.appdeploys.equipmentControlSystem.models.Users;
-
-import java.util.Optional;
-
-@Repository
-public interface UsersRepository extends JpaRepository<Users, Long> {
-    Optional<Users> findByEmail(String email);
-    Optional<Users> findByName(String name);
-}
