--- src/main/java/work/appdeploys/equipmentControlSystem/services/SchoolService.java
@@ -1,12 +1,0 @@
-package work.appdeploys.equipmentControlSystem.services;
-
-import work.appdeploys.equipmentControlSystem.models.dtos.SchoolDto;
-
-public interface SchoolService {
-
-    SchoolDto save(SchoolDto schoolDto);
-
-    void delete(Long id);
-
-    SchoolDto update(SchoolDto schoolDto);
-}
