--- src/main/java/work/appdeploys/equipmentControlSystem/services/UsersService.java
@@ -1,12 +1,0 @@
-package work.appdeploys.equipmentControlSystem.services;
-
-import work.appdeploys.equipmentControlSystem.models.dtos.UsersDto;
-
-public interface UsersService {
-
-    UsersDto save(UsersDto usersDto);
-
-    void delete(Long id);
-
-    UsersDto update(UsersDto usersDto);
-}
