+++ src/test/java/work/appDeploys/equipmentControlSystem/EquipmentControlSystemApplicationTests.java
@@ -1,0 +1,13 @@
+package work.appDeploys.equipmentControlSystem;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EquipmentControlSystemApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
