--- src/test/java/work/appdeploys/equipmentControlSystem/EquipmentControlSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package work.appdeploys.equipmentControlSystem;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EquipmentControlSystemApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
